+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.3.4] - 2019-02-13
+
 ### Changed
+* **kerberoast** action now has /domain and /dc like **asreproast** action
+* **kerberoast** and **asreproast** now properly work over domain trusts
 * **triage** command now works for the current non-elevated user, outputting current LUID as well
 * Current LUID output also added for non-elevated **dump** and **klist** commands
+* Added Opsec section in README.md
 
 
 ## [1.3.3] - 2019-02-11